+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/templating/mail/model/impl/EmailAttachmentImpl.java
@@ -19,14 +19,7 @@ package it.ozimov.springboot.templating.mail.model.impl;
 
 import it.ozimov.springboot.templating.mail.model.EmailAttachment;
 import it.ozimov.springboot.templating.mail.utils.TikaDetector;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
+import lombok.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.MediaType;
@@ -38,9 +31,8 @@ import java.io.InputStream;
 import static java.util.Optional.ofNullable;
 
 /**
- * Plain text email.
+ * Email attachment.
  *
- * @author rtrunfio
  */
 @NoArgsConstructor
 @RequiredArgsConstructor
@@ -54,14 +46,12 @@ public class EmailAttachmentImpl implements EmailAttachment {
     private static final long serialVersionUID = -3307831714212032363L;
 
     @Getter
-    private
     @NonNull
-    String attachmentName;
+    private String attachmentName;
 
     @Getter
-    private
     @NonNull
-    byte[] attachmentData;
+    private byte[] attachmentData;
 
     private MediaType mediaType;