+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/templating/mail/model/impl/EmailImpl.java
@@ -57,13 +57,11 @@ public class EmailImpl implements Email, Serializable {
 
     private Collection<InternetAddress> bcc;
 
-    private
     @NonNull
-    String subject;
+    private String subject;
 
-    private
     @NonNull
-    String body;
+    private String body;
 
     private Collection<EmailAttachmentImpl> attachments;
 
@@ -73,6 +71,8 @@ public class EmailImpl implements Email, Serializable {
 
     private Date sentAt;
 
+    private InternetAddress receiptTo;
+
+    private InternetAddress depositionNotificationTo;
+
 }
-
-