+++ CHANGELOG.md
@@ -4,6 +4,26 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.3.7] - 2017-01-09
+### Fixed
+- Support to email attachment has been fixed by removing the content type of attachments (issue 15)
+
+## [0.3.7] - 2017-01-09
+### Fixed
+- Support to email attachment has been fixed by removing the content type of attachments (issue 15)
+### Changed
+- Removed Apache Tika dependency
+
+## [0.3.6] - 2017-01-05
+### Changed
+- Using Spring Boot _1.4.3.RELEASE_
+### Fixed
+- Support to email attachment has been fixed (issue 15)
+
+## [0.3.5] - 2017-01-04
+### Added
+- Added receipt notification in email header
+
 ## [0.3.4] - 2016-07-05
 ### Changed
 - Using Spring Boot _1.3.6.RELEASE_